% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bound}
\alias{bound}
\title{Bounds for Q and g}
\usage{
bound(x, bds)
}
\arguments{
\item{x}{observed values of either Q or g.}

\item{bds}{list containing the upper and lower bound for x.}
}
\description{
Function to bound Q and g estimates with user specified bounds.
}
