% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_init.R
\name{cv_split}
\alias{cv_split}
\title{Split-specific predictions}
\usage{
cv_split(fold, data, estQ, estg)
}
\arguments{
\item{fold}{one of the folds from the folds object.}

\item{data}{data.frame object containing the time series with relevant time ordering.}

\item{estQ}{object of class \code{initEst} with \code{sl3} results for Q.}

\item{estg}{object of class \code{initEst} with \code{sl3} results for g.}
}
\value{
An object of class \code{tstmle}.
\describe{
\item{QAW}{Split-specific prediction of Q.}
\item{Q1W}{Split-specific prediction of Q(1,W).}
\item{Q0W}{Split-specific prediction of Q(0,W).}
\item{pA1}{Split-specific prediction of g.}
\item{B}{Split-specific prediction of the blip.}
\item{Rule}{Split-specific prediction of the maximizing rule.}
\item{K}{Split-specific absolute value of the blip.}
}
}
\description{
Function to estimate split-specific predictions for Q and g, and relevant parts for the
optimal treatment regime.
}
