% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_init.R
\name{cv_split_blip}
\alias{cv_split_blip}
\title{Split-specific blip predictions}
\usage{
cv_split_blip(fold, estSplit, data, SL.library, outcome_type)
}
\arguments{
\item{fold}{one of the folds from the folds object.}

\item{estSplit}{result of the cross-validated \code{cv_split} function.}

\item{data}{data.frame object containing the time series with C_o as the relevant covariates.}

\item{SL.library}{list of \code{sl3} algorithms for the fit of the blip function.}
}
\value{
An object of class \code{tstmle}.
\describe{
\item{fullPred}{Split-specific blip with full V-fold cross-validation and Super Learner results.}
\item{cvPred}{Split-specific blip with prediction for only a specific fold.}
\item{valSet}{Validation samples used for each fold.}
\item{B}{Estimated blip for the validation samples.}
}
}
\description{
Function to estimate split-specific predictions for the blip function.
}
