% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estBlip}
\alias{estBlip}
\title{Wrapper for split-specific Blip}
\usage{
estBlip(folds, estSplt, Q, blip_library)
}
\arguments{
\item{folds}{user-specified list of folds- it should correspond to an element of \code{origami}.}

\item{estSplt}{result object of \code{estSplit}.}

\item{Q}{data.frame containg the relevant outcome and covariates for estimating Q.}

\item{blip_library}{list of \code{sl3} algorithms for the fit of the blip function.}
}
\description{
Function to make split-specific blip estimation more streamlined.
}
