% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estSplit}
\alias{estSplit}
\title{Wrapper for split-specific predictions}
\usage{
estSplit(folds, Q, estQ, estg)
}
\arguments{
\item{folds}{user-specified list of folds- it should correspond to an element of \code{origami}.}

\item{Q}{data.frame containg the relevant outcome and covariates for estimating Q.}

\item{estQ}{Q result of \code{initEst} format.}

\item{estg}{g result of \code{initEst} format.}
}
\description{
Function to make split-specific predictions more streamlined.
}
