% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_vals}
\alias{extract_vals}
\title{Extract validation samples}
\usage{
extract_vals(folds, split_preds)
}
\arguments{
\item{folds}{user-specified list of folds- it should correspond to an element of \code{origami}.}

\item{split_preds}{Cross-validated result of \code{cv_split}.}
}
\description{
Function to extract validation samples from prediction on all samples for use in CV-TMLE.
}
