% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPT_estimate_tmle.R
\name{fluctuate}
\alias{fluctuate}
\title{Fluctuate function}
\usage{
fluctuate(tmledata, flucmod, subset = seq_len(nrow(tmledata)))
}
\arguments{
\item{tmledata}{\code{data.frame} containing all observed values for the A and Y node,
and either initial or updated estimates for g and Q.}

\item{flucmod}{fluctuation model used.}

\item{subset}{subset of the data used for fluctuation.}
}
\description{
Function to estimate logistic parametric submodel and get updated estmate logistic fluctuation.
}
