% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_init.R
\name{initEst}
\alias{initEst}
\title{Initial Estimation with sl3 and specified C_{o(t)} dimension}
\usage{
initEst(Y, X, folds = NULL, SL.library, outcome_type = "binomial")
}
\arguments{
\item{Y}{data.frame object containing the outcome variable.}

\item{X}{data.frame object containing the relevant covariates.}

\item{folds}{user-specified list of folds- it should correspond to an element of \code{origami}.}

\item{SL.library}{list of \code{sl3} algorithms to be used for estimation.}
}
\value{
An object of class \code{tstmle}.
\describe{
\item{cvRisk}{CV Risk for each algorithm and Super Learner.}
\item{valY}{Prediction based on the estimated SL fit.}
\item{SL.library}{list of \code{sl3} algorithms to be used for estimation.}
\item{folds}{user-specified list of folds- it should correspond to an element of \code{origami}.}
\item{fullFit}{Fit on all samples.}
\item{cvFit}{CV fit.}
}
}
\description{
This function relies on the stacked ensemble learner in order to estimate relevant
parts of the likelihood as guided by the efficient influence curve.
(best used for short-term dependence).
}
