% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_init.R
\name{initEst_setA}
\alias{initEst_setA}
\title{Fit Q with specified exposure}
\usage{
initEst_setA(data, fit, setA)
}
\arguments{
\item{data}{data.frame object containing the time series with relevant time ordering.}

\item{fit}{fit object from \code{initEst}.}

\item{setA}{specify the value each A node should be set to.}
}
\value{
An object of class \code{tstmle}.
\describe{
\item{valY}{Prediction based on the input \code{fit} object.}
}
}
\description{
Function to make estimation of Q with specified exposure easier.
}
