% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_init.R
\name{initFrame}
\alias{initFrame}
\title{Make initial dataframe}
\usage{
initFrame(data, Cy, Ca)
}
\arguments{
\item{data}{data.frame object containing the time series with relevant time ordering.}

\item{Cy}{numeric specifying possible Markov order for Y nodes.}

\item{Ca}{numeric specifying possible Markov order for A nodes.}
}
\value{
An object of class \code{tstmle}.
\describe{
\item{Y}{Observed outcome with relevant lagged time-points as specified by Cy.}
\item{A}{FObserved exposure with relevant lagged time-points as specified by Ca.}
\item{Ca}{Input Markov order for A.}
\item{Cy}{Input Markov order for Y.}
\item{data}{Input data.}
}
}
\description{
Function to create the initial data.frame \code{sl3} would expect with i.i.d data.
}
