% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_stack}
\alias{make_stack}
\title{Make a sl3 stack}
\usage{
make_stack(learner_lists)
}
\arguments{
\item{learner_list}{list of algorithms, possibly with additional parameters.}
}
\description{
Function to make passing algorithms to sl3 easier.
}
