% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPT_estimate_tmle.R
\name{OPT_estimate}
\alias{OPT_estimate}
\title{Estimate function for the context-specific parameter under user-specified rule}
\usage{
OPT_estimate(tmledata)
}
\arguments{
\item{tmledata}{\code{data.frame} containing all observed values for the A and Y node,
and either initial or updated estimates for g and Q.}
}
\description{
Function to estimate the context-specific parameter under user-specified rule and
coresponding relevant parts of the influence curve for the single intervention.
}
