% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPT_estimate_tmle.R
\name{OPT_update}
\alias{OPT_update}
\title{Update function for the context-specific parameter under user-specified rule}
\usage{
OPT_update(data, Qbounds)
}
\arguments{
\item{data}{\code{data.frame} containing all observed values for the A and Y node,
and either initial or updated estimates for g and Q.}

\item{Qbounds}{bounds for the Q estimates.}
}
\description{
Function to update the Q part of the likelihood using the specified fluctuation model for the
context-specific parameter under user-specified rule.
}
