% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{plot_OPT}
\alias{plot_OPT}
\title{Plot estimates for the optimal treatment regime parameter}
\usage{
plot_OPT(res)
}
\arguments{
\item{res}{Result of \code{tstmleOPT} function.}
}
\description{
Function to plot the estimated Phi for the optimal treatment regime,
and the relevant confidence intervals for each of the rules considered.
}
