% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{plot_ts}
\alias{plot_ts}
\title{Plot for a single time-series}
\usage{
plot_ts(data, plot = "xts")
}
\arguments{
\item{data}{data.frame object containing the time series with relevant time ordering.}

\item{plot}{create a plot of type "xts" or "pa".}
}
\description{
Function to plot the simulated or actual time-series.
}
