% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPT_estimate_tmle.R
\name{ruletmle}
\alias{ruletmle}
\title{Prepare for Optimal Individualized Treatment Regime TMLE Procedure}
\usage{
ruletmle(obsA, obsY, pA1, Q0W, Q1W, ruleA, maxIter = 1000,
  Qbounds = c(1e-04, 1 - 1e-04))
}
\arguments{
\item{obsA}{Observed exposure (A).}

\item{obsY}{Observed outcome (Y).}

\item{pA1}{Estimate of g, or P(A=1) obtained by calling \code{estSplit}.}

\item{Q0W}{Estimate of Q(0,W), or E(Y|A=0,W) obtained by calling \code{estSplit}.}

\item{Q1W}{Estimate of Q(1,W), or E(Y|A=1,W) obtained by calling \code{estSplit}.}

\item{ruleA}{user-specified rule for exposure under which we want to estimate the context-specific
mean of the outcome.}

\item{maxIter}{maximum number of iterations for the iterative TMLE.}

\item{Qbounds}{bounds for the Q estimates.}
}
\value{
An object of class \code{tstmle}.
\describe{
\item{tmlePsi}{Context-specific mean of the outcome under user-specified \code{ruleA} estimated
using TMLE methodology.}
\item{tmleSD}{Standard deviation of the context-specific mean of the outcome under
user-specified \code{ruleA} estimated using TMLE methodology.}
\item{tmleCI}{Confidence interval of the context-specific mean of the outcome under
user-specified \code{ruleA} estimated using TMLE methodology.}
\item{IC}{Influence curve for the context-specific parameter under user-specified \code{ruleA}.}
\item{rule}{Used rule for the exposure.}
\item{steps}{Number of steps until convergence of the iterative TMLE.}
\item{initialData}{Initial estimates of g and Q, and observed A and Y.}
\item{tmleData}{Final updates estimates of g, Q and clever covariates.}
\item{all}{Full results of \code{tmleOPT}.}
}
}
\description{
This function performs all the necessary steps in order to call the main TMLE function
for the optimal individualized treatment regime, \code{tmleOPT}. It also performs all the
calculations necessary to obtain ineference on the parameter.
}
