% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI_estimate_tmle.R
\name{SI_estimate}
\alias{SI_estimate}
\title{Estimate function}
\usage{
SI_estimate(tmledata)
}
\arguments{
\item{tmledata}{\code{data.frame} containing all observed values for the A and Y node,
estimate of g, Q, as well as Q evaluated when A=1 and A=0.}
}
\description{
Function to estimate the ATE parameter and coresponding relevant parts of the influence curve
for the single intervention contex-specific parameter.
}
