% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI_estimate_tmle.R
\name{sitmleEst}
\alias{sitmleEst}
\title{Main TMLE Calculations}
\usage{
sitmleEst()
}
\arguments{
\item{Y}{data.frame object containing the outcome variable.}
}
\value{
An object of class \code{tstmle}.
\describe{
\item{fitW}{Fit object for W part of the likelihood.}
}
}
\description{
This function runs the TMLE and computes the estimate.
}
