% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sl3.fit}
\alias{sl3.fit}
\title{Fit sl3 Super Learner}
\usage{
sl3.fit(task, SL.library)
}
\arguments{
\item{task}{task object of \code{sl3::make_sl3_Task} form.}

\item{SL.library}{list of \code{sl3} algorithms to be used for estimation. For the list of available
learners for time-series, see \code{sl3::sl3_list_learners(c("timeseries"))}.}
}
\description{
Function to make Super Learning with \code{sl3} easier.
}
