% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_ts_ATE.R
\name{tmle3_Spec_ts_ATE}
\alias{tmle3_Spec_ts_ATE}
\title{Defines a TML Estimator (except for the data)}
\description{
Defines a TML Estimator (except for the data)

Defines a TML Estimator (except for the data)
}
\section{Super class}{
\code{\link[tmle3:tmle3_Spec]{tmle3::tmle3_Spec}} -> \code{tmle3_Spec_ts_ATE}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-tmle3_Spec_ts_ATE-new}{\code{tmle3_Spec_ts_ATE$new()}}
\item \href{#method-tmle3_Spec_ts_ATE-make_tmle_task}{\code{tmle3_Spec_ts_ATE$make_tmle_task()}}
\item \href{#method-tmle3_Spec_ts_ATE-add_summary}{\code{tmle3_Spec_ts_ATE$add_summary()}}
\item \href{#method-tmle3_Spec_ts_ATE-make_params}{\code{tmle3_Spec_ts_ATE$make_params()}}
\item \href{#method-tmle3_Spec_ts_ATE-clone}{\code{tmle3_Spec_ts_ATE$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tmle3" data-topic="tmle3_Spec" data-id="make_initial_likelihood"><a href='../../tmle3/html/tmle3_Spec.html#method-tmle3_Spec-make_initial_likelihood'><code>tmle3::tmle3_Spec$make_initial_likelihood()</code></a></li>
<li><span class="pkg-link" data-pkg="tmle3" data-topic="tmle3_Spec" data-id="make_targeted_likelihood"><a href='../../tmle3/html/tmle3_Spec.html#method-tmle3_Spec-make_targeted_likelihood'><code>tmle3::tmle3_Spec$make_targeted_likelihood()</code></a></li>
<li><span class="pkg-link" data-pkg="tmle3" data-topic="tmle3_Spec" data-id="make_updater"><a href='../../tmle3/html/tmle3_Spec.html#method-tmle3_Spec-make_updater'><code>tmle3::tmle3_Spec$make_updater()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmle3_Spec_ts_ATE-new"></a>}}
\if{latex}{\out{\hypertarget{method-tmle3_Spec_ts_ATE-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_ts_ATE$new(treatment_level, control_level, Co, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmle3_Spec_ts_ATE-make_tmle_task"></a>}}
\if{latex}{\out{\hypertarget{method-tmle3_Spec_ts_ATE-make_tmle_task}{}}}
\subsection{Method \code{make_tmle_task()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_ts_ATE$make_tmle_task(data, node_list, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmle3_Spec_ts_ATE-add_summary"></a>}}
\if{latex}{\out{\hypertarget{method-tmle3_Spec_ts_ATE-add_summary}{}}}
\subsection{Method \code{add_summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_ts_ATE$add_summary(Co, W, A, Y, X, combine = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmle3_Spec_ts_ATE-make_params"></a>}}
\if{latex}{\out{\hypertarget{method-tmle3_Spec_ts_ATE-make_params}{}}}
\subsection{Method \code{make_params()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_ts_ATE$make_params(tmle_task, likelihood)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmle3_Spec_ts_ATE-clone"></a>}}
\if{latex}{\out{\hypertarget{method-tmle3_Spec_ts_ATE-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_ts_ATE$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
