% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPT_estimate_tmle.R
\name{tmleOPT}
\alias{tmleOPT}
\title{Main TMLE Calculations for the Optimal Individualized Treatment Regime Parameter}
\usage{
tmleOPT(tmledata, maxIter = 1000, Qbounds = c(1e-04, 1 - 1e-04))
}
\arguments{
\item{tmledata}{\code{data.frame} containing all observed values for the A and Y node,
estimate of g, Q, as well as Q evaluated when A=1 and A=0.}

\item{maxIter}{maximum number of iterations for the iterative TMLE.}

\item{Qbounds}{bounds for the Q estimates.}
}
\value{
An object of class \code{tstmle}.
\describe{
\item{tmledata}{Final updates estimates of g, Q and clever covariates.}
\item{psi}{Average treatment effect estimated using TMLE.}
\item{steps}{Number of steps until convergence of the iterative TMLE.}
\item{IC}{Influence function.}
\item{ED}{Mean of the final influence function.}
\item{ED2}{Mean of the squared final influence function.}
}
}
\description{
This function performs all the main TMLE calculations for the Context-Specific
Optimal Individualized Treatment Regime Parameter.
}
