% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI_estimate_tmle.R
\name{tmleSI}
\alias{tmleSI}
\title{Main TMLE Calculations for the Single Intervention Context-Specific ATE Paramater}
\usage{
tmleSI(tmledata, tol = 10^-3, maxIter = 1000, gbounds = gbounds,
  Qbounds = Qbounds)
}
\arguments{
\item{tmledata}{\code{data.frame} containing all observed values for the A and Y node,
estimate of g, Q, as well as Q evaluated when A=1 and A=0.}

\item{tol}{tolerance parameter for the iterative TMLE.}

\item{maxIter}{maximum number of iterations for the iterative TMLE.}

\item{gbounds}{bounds for the q estimates.}

\item{Qbounds}{bounds for the Q estimates.}
}
\value{
An object of class \code{tstmle}.
\describe{
\item{tmledata}{Final updates estimates of g, Q and clever covariates.}
\item{psi}{Average treatment effect estimated using TMLE.}
\item{steps}{Number of steps until convergence of the iterative TMLE.}
\item{IC}{Influence function.}
\item{ED}{Mean of the final influence function.}
\item{ED2}{Mean of the squared final influence function.}
}
}
\description{
This function performs all the main TMLE calculations for the single intervention
Context-Specific ATE Paramater.
}
