% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstmle3.R
\name{tstmle3}
\alias{tstmle3}
\title{Context-specific causal effect of single-time point intervention}
\usage{
tstmle3(
  data,
  node_list,
  parameter = "ATE",
  learner_list,
  Co = 5,
  cvtmle = FALSE,
  fold_fn = "folds_rolling_origin",
  first_window = 100,
  validation_size = 50,
  gap = 0,
  batch = 50,
  window_size = 100
)
}
\arguments{
\item{data}{data.frame object containing the data.}

\item{node_list}{node list reflecting the relationships between variables.}

\item{parameter}{target parameter for targeting. Default is the average over time
Context-Specific Average Treatment Effect.}

\item{learner_list}{learner list containing the learners used for the conditional expectation
of outcome and propensity score.}

\item{Co}{user-specified Markov order for the fixed dimensional summary measure.}

\item{cvtmle}{default is \code{FALSE}.}

\item{fold_fn}{cross-validation scheme, as defined by \code{origami}. See \code{?origami::fold_funs}
 for detailed explanations. For time-series, implemented cross-validation schemes are
\code{folds_rolling_origin} and \code{folds_rolling_window}.}

\item{first_window}{first window size used for training. Only relevant if the set cross-validation is
\code{folds_rolling_origin}.}

\item{validation_size}{number of time points used for validation.}

\item{gap}{number of time points between training and validation set.}

\item{batch}{number of time points added in the next next cross-validation fold.}

\item{window_size}{window size used for training. Only relevant if the set cross-validation is
\code{folds_rolling_window}.}
}
\value{
An object of class \code{tstmle}.
\describe{
\item{tmlePsi}{Average treatment effect estimated using TMLE.}
\item{iptwPsi}{Average treatment effect estimated using IPTW.}
\item{tmleSE}{Standard error for the TMLE estimated parameter.}
\item{tmleSD}{Standard deviation for the TMLE estimated parameter.}
\item{tmleCI}{Confidence Interval for the TMLE estimated parameter.}
\item{IC}{Influence function.}
\item{steps}{Number of steps until convergence of the iterative TMLE.}
\item{initialData}{Initial estimates of g, Q.}
\item{tmleData}{Final updates estimates of g, Q and clever covariates.}
\item{tmle_fit}{The full tmle3 fit object.}
}
}
\description{
This function estimates the causal effect of a single time-point intervention on the outcome
within the same time point. Here, time is defined as observing a single intervention and outcome,
and possibly multiple covariates. Intervention is imposed within each unit of time.
}
