% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI_tstmle.R
\name{tstmleMI}
\alias{tstmleMI}
\title{Context-specific causal effect of multiple time point interventions}
\usage{
tstmleMI(data, Co = NULL, block = 1, folds = NULL, V = 5,
  stratifyAY = TRUE, Q_library = list("Lrnr_mean", "Lrnr_glm_fast",
  "Lrnr_glmnet"), g_library = list("Lrnr_mean", "Lrnr_glm_fast",
  "Lrnr_glmnet"), gbounds = c(1e-04, 1 - 1e-04), Qbounds = c(1e-04, 1 -
  1e-04), maxIter = 1000, fold_fn = "folds_rolling_origin", window = 1,
  skip = 0)
}
\arguments{
\item{data}{data.frame object containing the time series with relevant time ordering.}

\item{Co}{numeric specifying possible Markov order for the O(t) nodes.}

\item{block}{size of the new O(t) block: multiple of the original O(t). Default of 1 will give the
original time-series data-frame.}

\item{folds}{user-specified list of folds- it should correspond to an element of \code{origami}.
In case it is not specified, it will defined internally.}

\item{V}{number of cross-validation folds used.}

\item{stratifyAY}{logical: should we stratify the cross-validation based on (A,Y) pairs}

\item{Q_library}{list of \code{sl3} algorithms for the fit of E(Y|A,Cy) (Q)}

\item{g_library}{list of \code{sl3} algorithms for the fit of P(A|Ca) (g)}

\item{gbounds}{bounds for the q estimates.}

\item{Qbounds}{bounds for the Q estimates.}

\item{maxIter}{maximum number of iterations for the iterative TMLE.}
}
\value{
An object of class \code{tstmleSI}.
\describe{
\item{tmlePsi}{Average treatment effect estimated using TMLE.}
\item{iptwPsi}{Average treatment effect estimated using IPTW.}
\item{tmleSD}{Standard deviation for the TMLE estimated parameter.}
\item{tmleCI}{Confidence Interval for the TMLE estimated parameter.}
\item{IC}{Influence function.}
\item{steps}{Number of steps until convergence of the iterative TMLE.}
\item{initialData}{Initial estimates of g, Q.}
\item{tmleData}{Final updates estimates of g, Q and clever covariates.}
}
}
\description{
This function estimates the causal effect of multiple time-point interventions in an ordered
longitudinal data structure within time unit t. Here, time is defined as observing multiple ordered
logitudinal data structures of user-specified dimension, with possibly multiple intervention and
outcome/covariate nodes within it. Time-block defined in this manner could be artificially created
for the purpose of estimation of a particular causal effect, of they could have a unique meaning such
as representing measurments over a day/week/cycle.
}
