% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPT_tstmle.R
\name{tstmleOPT}
\alias{tstmleOPT}
\title{Context-specific optimal treatment effect for a single time series}
\usage{
tstmleOPT(data, Co = TRUE, Cy = NULL, Ca = NULL, folds = NULL, V = 5,
  stratifyAY = TRUE, Q_library = list("Lrnr_mean", "Lrnr_glm_fast",
  "Lrnr_glmnet"), g_library = list("Lrnr_mean", "Lrnr_glm_fast",
  "Lrnr_glmnet"), blip_library = list("Lrnr_mean", "Lrnr_glm_fast",
  "Lrnr_glmnet", "Lrnr_randomForest", "Lrnr_xgboost"), gbounds = c(1e-04, 1 -
  1e-04), Qbounds = c(1e-04, 1 - 1e-04), maxIter = 1000)
}
\arguments{
\item{data}{data.frame object containing the time series with relevant time ordering.}

\item{Cy}{numeric specifying possible Markov order for Y nodes.}

\item{Ca}{numeric specifying possible Markov order for A nodes.}

\item{folds}{user-specified list of folds- it should correspond to an element of \code{origami}.
In case it is not specified, it will defined internally.}

\item{V}{number of cross-validation folds used.}

\item{stratifyAY}{logical: should we stratify the cross-validation based on (A,Y) pairs}

\item{Q_library}{list of \code{sl3} algorithms for the fit of E(Y|A,Cy) (Q)}

\item{g_library}{list of \code{sl3} algorithms for the fit of P(A|Ca) (g)}

\item{blip_library}{list of \code{sl3} algorithms for the fit of the blip function.}

\item{gbounds}{bounds for the q estimates.}

\item{Qbounds}{bounds for the Q estimates.}

\item{maxIter}{maximum number of iterations for the iterative TMLE.}
}
\value{
An object of class \code{tstmleOPT}.
\describe{
\item{tmlePsi}{Average treatment effect estimated using TMLE.}
\item{iptwPsi}{Average treatment effect estimated using IPTW.}
\item{tmleSD}{Standard deviation for the TMLE estimated parameter.}
\item{tmleCI}{Confidence Interval for the TMLE estimated parameter.}
\item{IC}{Influence function.}
\item{steps}{Number of steps until convergence of the iterative TMLE.}
\item{initialData}{Initial estimates of g, Q.}
\item{tmleData}{Final updates estimates of g, Q and clever covariates.}
}
}
\description{
This function estimates the optimal individualized treatment effect for a single time series.
}
