% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI_tstmle.R
\name{tstmleSI}
\alias{tstmleSI}
\title{Context-specific causal effect of single-time point intervention}
\usage{
tstmleSI(data, Cy = NULL, Ca = NULL, folds = NULL, V = 5,
  stratifyAY = TRUE, Q_library = list("Lrnr_mean", "Lrnr_glm_fast",
  "Lrnr_glmnet"), g_library = list("Lrnr_mean", "Lrnr_glm_fast",
  "Lrnr_glmnet"), Q_SLlibrary = list("Lrnr_mean", "Lrnr_arima",
  "Lrnr_expSmooth"), g_SLlibrary = list("Lrnr_mean", "Lrnr_arima",
  "Lrnr_expSmooth"))
}
\arguments{
\item{data}{data.frame object containing the time series with relevant time ordering.}

\item{Cy}{numeric specifying possible Markov order for Y nodes.}

\item{Ca}{numeric specifying possible Markov order for A nodes.}

\item{folds}{user-specified list of folds- it should correspond to an element of \code{origami}.
In case it is not specified, it will defined internally.}

\item{V}{number of cross-validation folds used.}

\item{stratifyAY}{logical: should we stratify the cross-validation based on (A,Y) pairs}

\item{Q_library}{list of \code{sl3} algorithms for the fit of E(Y|A,Cy) (Q)}

\item{g_library}{list of \code{sl3} algorithms for the fit of P(A|Ca) (g)}

\item{Q_SLlibrary}{list of \code{sl3} algorithms to be used for estimation of E(Y|A,Cy) (Q).
For the list of available learners for time-series, see \code{sl3::sl3_list_learners()}.}

\item{q_SLlibrary}{list of \code{sl3} algorithms to be used for estimation of P(A|Ca) (g).
For the list of available learners for time-series, see \code{sl3::sl3_list_learners()}.}

\item{fold_fn}{cross-validation scheme, as defined by \code{origami}. See \code{?origami::fold_funs}
for detailed explanations. For time-series, implemented cross-validation schemes are
\code{folds_rolling_origin} and \code{folds_rolling_window}.}

\item{window}{in case \code{fold_fn} was set to \code{folds_rolling_window}, specify the
number of observations in each training sample.}

\item{skip}{in case the time-series considered is very long, it is possible there will be many
folds to consider. This parameter allows for few nodes to be skipped. Default is 0, which
corresonds to no nodes skipped.}
}
\value{
An object of class \code{tstmle}.
\describe{
\item{fitW}{Fit object for W part of the likelihood.}
}
}
\description{
This function estimates the causal effect of a single time-point intervention on the outcome
within the same time point. Here, time is defined as observing a single intervention and outcome,
and possibly multiple covariates. Intervention is impossed within each unit of time.
}
