% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bound}
\alias{bound}
\title{Bounds for Q and g}
\usage{
bound(x, bds)
}
\arguments{
\item{learner_list}{list of algorithms, possibly with additional parameters.}
}
\description{
Function to bound Q and g estimates
}
