% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_val}
\alias{extract_val}
\title{Extract validation samples}
\usage{
extract_val(fold, split_preds)
}
\arguments{
\item{fold}{one fold from a list of folds.}

\item{split_preds}{Cross-validated result of \code{cv_split}.}
}
\description{
Function to extract the validation sets from split based predictions for use in CV-TMLE.
}
