% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPT_estimate_tmle.R, R/SI_estimate_tmle.R
\name{fluctuate}
\alias{fluctuate}
\alias{fluctuate}
\title{Fluctuate function}
\usage{
fluctuate(tmledata, flucmod, subset = seq_len(nrow(tmledata)))

fluctuate(tmledata, flucmod, subset = seq_len(nrow(tmledata)))
}
\arguments{
\item{tmledata}{\code{data.frame} containing all observed values for the A and Y node,
estimate of g, Q, as well as Q evaluated when A=1 and A=0.}

\item{flucmod}{fluctuation model used.}

\item{subset}{subset of the data used for fluctuation.}

\item{tmledata}{\code{data.frame} containing all observed values for the A and Y node,
estimate of g, Q, as well as Q evaluated when A=1 and A=0.}

\item{flucmod}{fluctuation model used.}

\item{subset}{subset of the data used for fluctuation.}
}
\description{
Function to estimate logistic parametric submodel and get updated estmate logistic fluctuation

Function to estimate logistic parametric submodel and get updated estmate logistic fluctuation
}
