% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_init.R, R/initEst.R
\name{initEst}
\alias{initEst}
\alias{initEst}
\title{Initial Estimation with sl3 and specified C_{o(t)} dimension}
\usage{
initEst(Y, X)

initEst(Y, X)
}
\arguments{
\item{Y}{data.frame object containing the outcome variable.}

\item{X}{data.frame object containing the relevant covariates.}

\item{folds}{user-specified list of folds- it should correspond to an element of \code{origami}.}

\item{SL.library}{list of \code{sl3} algorithms to be used for estimation.}

\item{Y}{data.frame object containing the outcome variable.}

\item{X}{data.frame object containing the relevant covariates.}

\item{folds}{user-specified list of folds- it should correspond to an element of \code{origami}.
In case it is not specified, it will defined internally.}

\item{fold_fn}{cross-validation scheme, as defined by \code{origami}. See \code{origami::fold_funs}
for detailed explanations. For time-series, implemented cross-validation schemes are
\code{folds_rolling_origin} and \code{folds_rolling_window}.}

\item{SL.library}{list of \code{sl3} algorithms to be used for estimation. For the list of available
learners for time-series, see \code{sl3::sl3_list_learners(c("timeseries"))}.}
}
\value{
An object of class \code{tstmle}.
\describe{
\item{cvRisk}{CV Risk for each algorithm and Super Learner.}
\item{valY}{Prediction based on the estimated SL fit.}
\item{SL.library}{list of \code{sl3} algorithms to be used for estimation.}
\item{folds}{user-specified list of folds- it should correspond to an element of \code{origami}.}
\item{fullFit}{Fit on all samples.}
\item{cvFit}{CV fit.}
}

An object of class \code{tstmle}.
\describe{
\item{fitW}{Fit object for W part of the likelihood.}
}
}
\description{
This function relies on the stacked ensemble learner in order to estimate relevant
parts of the likelihood as guided by the efficient influence curve.
(best used for short-term dependence).

This function relies on the stacked ensemble learner in order to estimate relevant
parts of the likelihood as guided by the efficient influence curve.
(best used for short-term dependence).
}
