% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_init.R
\name{initFrame_mi}
\alias{initFrame_mi}
\title{Make initial dataframe for multiple intervention setting}
\usage{
initFrame_mi(data, Co, block = 1)
}
\arguments{
\item{data}{data.frame object containing the time series with relevant time ordering.}

\item{Co}{numeric specifying possible Markov order for the O(t) nodes.}

\item{block}{size of the new O(t) block: multiple of the original O(t). Default of 1 will give the
original time-series data-frame.}
}
\value{
An object of class \code{tstmle}.
\describe{
\item{Y}{data set containing the \code{sl3} ready input for the output and relevant covariates.}
\item{A}{data set containing the \code{sl3} ready input for all the exposures and relevant covariates.}
\item{final}{Final dataset with each list component representing a data set with targeted node
list of covariates as specified by Co and batch t.}
\item{original}{Final dataset with each list component representing a data set with targeted node
list of covariates as specified by Co.}
\item{step}{Number of nodes in an initial batch.}
\item{card}{Cardinality of the newly created block.}
}
}
\description{
Function to create the initial data.frame \code{sl3} would expect with i.i.d data.
The newly created data will rely on the specified Cl and Ca, as well as on the size of the O(t)
block.
}
