% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI_tstmle.R
\name{tstmleSI}
\alias{tstmleSI}
\title{Context-specific causal effect of single-time point intervention}
\usage{
tstmleSI(data, Cy = NULL, Ca = NULL, folds = NULL, V = 5,
  stratifyAY = TRUE, Q_library = list("Lrnr_mean", "Lrnr_glm_fast",
  "Lrnr_glmnet"), g_library = list("Lrnr_mean", "Lrnr_glm_fast",
  "Lrnr_glmnet"), Q_SLlibrary = list("Lrnr_mean", "Lrnr_arima",
  "Lrnr_expSmooth"), g_SLlibrary = list("Lrnr_mean", "Lrnr_arima",
  "Lrnr_expSmooth"), gbounds = c(1e-04, 1 - 1e-04), Qbounds = c(1e-04, 1 -
  1e-04), tol = 10^-3, maxIter = 1000)
}
\arguments{
\item{data}{data.frame object containing the time series with relevant time ordering.}

\item{Cy}{numeric specifying possible Markov order for Y nodes.}

\item{Ca}{numeric specifying possible Markov order for A nodes.}

\item{folds}{user-specified list of folds- it should correspond to an element of \code{origami}.
In case it is not specified, it will defined internally.}

\item{V}{number of cross-validation folds used.}

\item{stratifyAY}{logical: should we stratify the cross-validation based on (A,Y) pairs}

\item{Q_library}{list of \code{sl3} algorithms for the fit of E(Y|A,Cy) (Q)}

\item{g_library}{list of \code{sl3} algorithms for the fit of P(A|Ca) (g)}

\item{Q_SLlibrary}{list of \code{sl3} algorithms to be used for estimation of E(Y|A,Cy) (Q).
For the list of available learners for time-series, see \code{sl3::sl3_list_learners()}.}

\item{gbounds}{bounds for the q estimates.}

\item{Qbounds}{bounds for the Q estimates.}

\item{tol}{tolerance parameter for the iterative TMLE.}

\item{maxIter}{maximum number of iterations for the iterative TMLE.}

\item{q_SLlibrary}{list of \code{sl3} algorithms to be used for estimation of P(A|Ca) (g).
For the list of available learners for time-series, see \code{sl3::sl3_list_learners()}.}
}
\value{
An object of class \code{tstmle}.
\describe{
\item{tmlePsi}{Average treatment effect estimated using TMLE.}
\item{iptwPsi}{Average treatment effect estimated using IPTW.}
\item{tmleSD}{Standard deviation for the TMLE estimated parameter.}
\item{tmleCI}{Confidence Interval for the TMLE estimated parameter.}
\item{IC}{Influence function.}
\item{steps}{Number of steps until convergence of the iterative TMLE.}
\item{initialData}{Initial estimates of g, Q.}
\item{tmleData}{Final updates estimates of g, Q and clever covariates.}
}
}
\description{
This function estimates the causal effect of a single time-point intervention on the outcome
within the same time point. Here, time is defined as observing a single intervention and outcome,
and possibly multiple covariates. Intervention is impossed within each unit of time.
}
